requirements.txt
absl-py==2.4.0
accelerate==0.27.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
apache-tvm-ffi==0.1.8.post2
astor==0.8.1
async-timeout==5.0.1
attributedict==0.3.0
attrs==25.4.0
blake3==1.0.8
blessings==1.7
cachetools==7.0.0
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==7.13.3
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cuda-python==13.1.1
cupy-cuda12x==13.6.0
DataProperty==1.1.0
datasets==4.5.0
deepdiff==8.6.1
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
evaluate==0.4.6
exceptiongroup==1.3.1
fastapi==0.128.5
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
fastrlock==0.8.3
filelock==3.20.3
flashinfer-python==0.6.1
frozenlist==1.8.0
fsspec==2025.10.0
gguf==0.17.1
grpcio==1.78.0
grpcio-reflection==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
ijson==3.4.0.post0
inspecta==0.1.3
interegular==0.3.3
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lm_eval==0.4.10
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mbstrdecoder==1.1.4
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.9.0
model-hosting-container-standards==0.1.13
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
ninja==1.13.0
nltk==3.9.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.3.5
nvidia-ml-py==13.590.48
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.17.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
orderly-set==5.5.0
outlines_core==0.2.11
packaging==26.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pathvalidate==3.3.1
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
portalocker==3.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.0
pybase64==1.4.3
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyproject-api==1.10.0
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
rootpath==0.1.1
rouge_score==0.1.2
rpds-py==0.30.0
sacrebleu==2.6.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.52.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlitedict==2.1.0
sse-starlette==3.2.0
starlette==0.52.1
supervisor==4.3.0
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
termcolor==3.3.0
texttable==1.7.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.4.0
torch==2.1.2+cu118
tox==4.34.1
tqdm==4.67.3
transformers==4.51.3
triton==2.1.0
typepy==1.3.4
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websockets==16.0
word2number==1.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
